Image Metadata Analysis/app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Pillow